requirements.txt
pyyaml>=6.0
pymavlink>=2.4.40
pyserial>=3.5
textual>=0.50.0
